+++ CHANGELOG.md
@@ -2,10 +2,21 @@
 
 ## 0.3.0 (\_\_\_)
 
+**BREAKING CHANGES**:
+
+- DateControl:
+  - Was
+    - changing `$options.format` & attr `[format]` related to all date-strings: attributes `initvalue/min/max`.
+    - default value `YYYY-MM-DD`
+  - Now
+    - changing `$options.format` & attr `[format]` related only to displayed text. All attributes must be pointed in universal format `YYYY-MM-DD`
+    - default value depends on user localization; see [locale](src/helpers/locale.ts)
+
 **Features**:
 
 - Added [helpers](README.md#helpers)
   - [`onScroll`](src/helpers/onScroll.ts)
+  - [`locale`](src/helpers/locale.ts)
 
 ## 0.2.0 (Dec 09, 2022)